Session bug fix
:x
